+++ apps/web/components/overlay/widgets/PrePostFlow.tsx
@@ -32,3 +32,4 @@ export { SupporterWallWidget } from "./SupporterWallWidget";
 export { ClipFlagWidget } from "./ClipFlagWidget";
 export { LowerThirdWidget } from "./LowerThirdWidget";
 export { StatsTickerWidget } from "./StatsTickerWidget";
+export { CountdownTimerWidget } from "./CountdownTimerWidget";